requirements.txt
pandas>=2.0.0
python-dotenv>=1.0.0
crewai>=0.1.0
crewai-tools>=0.1.0
openai>=1.0.0
langchain-openai>=0.0.1 
matplotlib>=3.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
